refactor: move example into output folder
