+++ csv_detective/__init__.py
@@ -1,4 +1,4 @@
 from .explore_csv import routine, routine_minio  # noqa
-from .create_example import create_example_csv_file  # noqa
+from .output.example import create_example_csv_file  # noqa
 
 __version__ = '0.7.5.dev'
